Move Forward as Attachment to app menu and fix eslint regex errors
- Remove "Forward as Attachment" from the per-message dropdown and
  add it as a Thread > Forward as Attachment menu item (all platforms)
- Register core:forward-as-attachment command handler in MessageList
- Fix no-useless-escape: use character class [/?<>\\:*|"] instead
  of individually escaping each character
- Fix no-control-regex: use \u0000-\u001f unicode escapes in eml-utils
